6.2.07c
Added lost pages which are gained every 20 lore you unlock and used to
unlock unavailable lore
Added the ability to unlock unavailable lore and get an idea of the
content of unavailable lore before unlocking.
Fixed PoI disapearing when the lore of that location unlocks.
Fixed list and display updating when unlocking lore